# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server celler server.",
    "title": "Nunu Example API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "paths": {
    "/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "用户登录注册",
        "parameters": [
          {
            "description": "params",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.Response"
            }
          }
        }
      }
    },
    "/send-code": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "发送验证码",
        "parameters": [
          {
            "description": "params",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.SendVerifyCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.Response"
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "获取用户信息",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.GetProfileResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "修改用户信息",
        "parameters": [
          {
            "description": "params",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.UpdateProfileRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.Response"
            }
          }
        }
      }
    },
    "/user/feedback/add": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户反馈模块"
        ],
        "summary": "新增用户反馈",
        "parameters": [
          {
            "description": "params",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.AddUserFeedbackRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/aphrodite-go_api_v1.Response"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "aphrodite-go_api_v1.AddUserFeedbackRequest": {
      "type": "object",
      "required": [
        "feedback"
      ],
      "properties": {
        "feedback": {
          "type": "string",
          "example": "banana"
        }
      }
    },
    "aphrodite-go_api_v1.GetProfileResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/aphrodite-go_api_v1.GetProfileResponseData"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "aphrodite-go_api_v1.GetProfileResponseData": {
      "type": "object",
      "required": [
        "phone"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "1234@gmail.com"
        },
        "nickname": {
          "type": "string",
          "example": "banana"
        },
        "phone": {
          "type": "string",
          "example": "13288888888"
        },
        "userCode": {
          "type": "string"
        },
        "userNo": {
          "type": "string"
        }
      }
    },
    "aphrodite-go_api_v1.LoginRequest": {
      "type": "object",
      "properties": {
        "openId": {
          "type": "string",
          "example": "123456"
        },
        "phone": {
          "type": "string",
          "example": "123456"
        },
        "verifyCode": {
          "type": "string",
          "example": "1234"
        }
      }
    },
    "aphrodite-go_api_v1.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "message": {
          "type": "string"
        }
      }
    },
    "aphrodite-go_api_v1.SendVerifyCodeRequest": {
      "type": "object",
      "required": [
        "phone"
      ],
      "properties": {
        "phone": {
          "type": "string",
          "example": "13288888888"
        }
      }
    },
    "aphrodite-go_api_v1.UpdateProfileRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "1234@gmail.com"
        },
        "newPassword": {
          "type": "string",
          "example": "banana"
        },
        "newPhone": {
          "type": "string",
          "example": "13288888888"
        },
        "nickname": {
          "type": "string",
          "example": "banana"
        },
        "oldPassword": {
          "type": "string",
          "example": "banana"
        },
        "oldPhone": {
          "type": "string",
          "example": "13288888888"
        },
        "verifyCode": {
          "type": "string",
          "example": "1234"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}